third_party/WebKit/LayoutTests/http/tests/inspector/resource-tree/resource-tree-write-to-open-document.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../resources-test.js"></script>
<script src="resource-tree-test.js"></script>

<script>
function loadIframe()
{
    var iframe = document.createElement("iframe");
    document.body.appendChild(iframe);
    iframe.src = "resources/dynamic-iframe.html";
}

function test()
{
    for (var eventName in WebInspector.ResourceTreeModel.EventTypes)
        WebInspector.resourceTreeModel.addEventListener(WebInspector.ResourceTreeModel.EventTypes[eventName], eventHandler.bind(this, eventName));

    var frames = [];
    function frameId(frame) {
        var frameId = frames.indexOf(frame) + 1;
        if (!frameId) {
            frames.push(frame);
            frameId = frames.length;
        }
        return frameId;
    }

    function eventHandler(eventName, event)
    {
        switch (eventName) {
        case "FrameAdded":
        case "FrameNavigated":
        case "FrameDetached":
        case "MainFrameNavigated":
            var frame = event.data;
            var parentFrame = frame.parentFrame;
            parentFrame = parentFrame ? (", parentFrameId: " + frameId(parentFrame)) : "";
            InspectorTest.addResult("    " + eventName + " id: " + frameId(frame) + parentFrame);
            break;
        default:
        }
    }

    InspectorTest.addConsoleSniffer(InspectorTest.runAfterPendingDispatches.bind(null, InspectorTest.completeTest));
    InspectorTest.evaluateInPage("loadIframe();");
}
</script>

</head>
<body onload="runTest()">
<p>
Tests that frame writing to open document is not reported as frame navigation.
<a href="https://code.google.com/p/chromium/issues/detail?id=315618">crbug/315618</a>
</p>
</body>
</html>
