third_party/WebKit/LayoutTests/http/tests/inspector/resource-tree/resource-tree-write-to-open-document-expected.txt

Tests that frame writing to open document is not reported as frame navigation. crbug/315618


    FrameAdded id: 2, parentFrameId: 1
    FrameNavigated id: 2, parentFrameId: 1
    FrameNavigated id: 2, parentFrameId: 1
    FrameAdded id: 3, parentFrameId: 2
    FrameNavigated id: 3, parentFrameId: 2
    FrameNavigated id: 3, parentFrameId: 2
